Rename net5 files, classes to net6
