+++ src/client/Microsoft.Identity.Client/PlatformsCommon/Factories/PlatformProxyFactory.cs
@@ -1,10 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
-using System;
 using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Internal.Logger;
+using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 
 namespace Microsoft.Identity.Client.PlatformsCommon.Factories
 {
@@ -23,7 +22,7 @@ namespace Microsoft.Identity.Client.PlatformsCommon.Factories
 #if NET_CORE
             return new Microsoft.Identity.Client.Platforms.netcore.NetCorePlatformProxy(finalLogger);
 #elif NET6_WIN
-            return new Microsoft.Identity.Client.Platforms.net5win.Net5WinPlatformProxy(finalLogger);
+            return new Microsoft.Identity.Client.Platforms.net6win.Net6WinPlatformProxy(finalLogger);
 #elif ANDROID
             return new Microsoft.Identity.Client.Platforms.Android.AndroidPlatformProxy(finalLogger);
 #elif iOS
